+++ datastore/couchDB/adminServiceDatastoreCouchDB.go
@@ -313,25 +313,14 @@ func (asd *AdminServiceDatastoreCouchDB) addTenantViewsToDB(dbName string) error
 
 	logger.Log.Debugf("Adding Tenant Views to DB %s", dbName)
 
-	db, err := getDatabase(createDBPathStr(asd.couchHost, dbName))
-	if err != nil {
-		return err
-	}
-
 	moDB, err := getDatabase(createDBPathStr(asd.couchHost, fmt.Sprintf("%s%s", dbName, monitoredObjectDBSuffix)))
 	if err != nil {
 		return err
 	}
 
-	// Store the sync checkpoint in CouchDB
-	for _, viewPayload := range generateTenantViews() {
-
-		viewDB := db
-
-		if viewPayload["_id"] == monitoredObjectCountByDomainIndex {
-			viewDB = moDB
-		}
-		_, _, err = storeDataInCouchDBWithQueryParams(viewPayload, "TenantView", viewDB, nil)
+	// Store the views related to Monitored Objects
+	for _, viewPayload := range getTenantViews() {
+		_, _, err = storeDataInCouchDBWithQueryParams(viewPayload, "TenantView", moDB, nil)
 		if err != nil {
 			return err
 		}
@@ -595,23 +584,6 @@ func (asd *AdminServiceDatastoreCouchDB) DeleteValidTypes() (*admmod.ValidTypes,
 	logger.Log.Debugf("Deleted %s: %v\n", admmod.ValidTypesStr, models.AsJSONString(existing))
 	return existing, nil
 
-}
-
-// Produces all of the views/indicies necessary for the Tenant DB
-func generateTenantViews() []map[string]interface{} {
-	result := make([]map[string]interface{}, 0)
-
-	monitoredObjectCountByDomain := map[string]interface{}{}
-	monitoredObjectCountByDomain["_id"] = monitoredObjectCountByDomainIndex
-	monitoredObjectCountByDomain["language"] = "javascript"
-	byDomain := map[string]interface{}{}
-	byDomain["map"] = "function(doc) {\n    if (doc.data && doc.data.datatype && doc.data.datatype === 'monitoredObject' && doc.data.domainSet) {\n      for (var i in doc.data.domainSet) {\n        emit(doc.data.domainSet[i], doc._id);\n      }\n    }\n}"
-	views := map[string]interface{}{}
-	views["byDomain"] = byDomain
-	monitoredObjectCountByDomain["views"] = views
-
-	logger.Log.Debug("Adding view for monitoredObjectCountByDomain")
-	return append(result, monitoredObjectCountByDomain)
 }
 
 func generateAdminViews() []map[string]interface{} {